feat: make bot bootable
The bot will now connect to both a PostgreSQL database and the Discord
API. You may need to copy the `.env.example` file over to a `.env`
file, and make changes to the environment variables when necessary.

Also introduces tracing as a logging framework.
Sentry may be added further down the line, if necessary.